+++ packages/graphql/src/schema/mutation.ts
@@ -662,6 +662,7 @@ export const Mutation = builder.mutationType({
           displayMode: t.arg({ required: false, type: QuestionDisplayMode }),
           hasSampleSolution: t.arg.boolean({ required: false }),
           hasAnswerFeedbacks: t.arg.boolean({ required: false }),
+          pointsMultiplier: t.arg.int({ required: false }),
           tags: t.arg.stringList({ required: false }),
           options: t.arg({
             type: OptionsChoicesInput,
@@ -686,6 +687,7 @@ export const Mutation = builder.mutationType({
           explanation: t.arg.string({ required: false }),
           hasSampleSolution: t.arg.boolean({ required: false }),
           hasAnswerFeedbacks: t.arg.boolean({ required: false }),
+          pointsMultiplier: t.arg.int({ required: false }),
           tags: t.arg.stringList({ required: false }),
           options: t.arg({
             type: OptionsNumericalInput,
@@ -710,6 +712,7 @@ export const Mutation = builder.mutationType({
           explanation: t.arg.string({ required: false }),
           hasSampleSolution: t.arg.boolean({ required: false }),
           hasAnswerFeedbacks: t.arg.boolean({ required: false }),
+          pointsMultiplier: t.arg.int({ required: false }),
           tags: t.arg.stringList({ required: false }),
           options: t.arg({
             type: OptionsFreeTextInput,
@@ -757,6 +760,50 @@ export const Mutation = builder.mutationType({
         },
         resolve(_, args, ctx) {
           return CourseService.changeCourseDates(args, ctx)
+        },
+      }),
+
+      publishLearningElement: asUser.field({
+        nullable: true,
+        type: LearningElement,
+        args: {
+          id: t.arg.string({ required: true }),
+        },
+        resolve(_, args, ctx) {
+          return LearningElementService.publishLearningElement(args, ctx)
+        },
+      }),
+
+      deleteLearningElement: asUser.field({
+        nullable: true,
+        type: LearningElement,
+        args: {
+          id: t.arg.string({ required: true }),
+        },
+        resolve(_, args, ctx) {
+          return LearningElementService.deleteLearningElement(args, ctx)
+        },
+      }),
+
+      publishMicroSession: asUser.field({
+        nullable: true,
+        type: MicroSession,
+        args: {
+          id: t.arg.string({ required: true }),
+        },
+        resolve(_, args, ctx) {
+          return MicroLearningService.publishMicroSession(args, ctx)
+        },
+      }),
+
+      deleteMicroSession: asUser.field({
+        nullable: true,
+        type: MicroSession,
+        args: {
+          id: t.arg.string({ required: true }),
+        },
+        resolve(_, args, ctx) {
+          return MicroLearningService.deleteMicroSession(args, ctx)
         },
       }),
     }