+++ CHANGELOG.md
@@ -2,6 +2,39 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.0.0-alpha.67](https://github.com/uzh-bf/klicker-uzh/compare/v3.0.0-alpha.66...v3.0.0-alpha.67) (2023-03-01)
+
+
+### Features
+
+* learning elements and micro sessions draft management and various improvements ([#3417](https://github.com/uzh-bf/klicker-uzh/issues/3417)) ([7c7fd1e](https://github.com/uzh-bf/klicker-uzh/commit/7c7fd1e815e6de6d176700614e6650a7d2ae0023))
+
+
+### Enhancements
+
+* **apps/frontend-manage:** add points multiplier to question forms ([#3425](https://github.com/uzh-bf/klicker-uzh/issues/3425)) ([79deb60](https://github.com/uzh-bf/klicker-uzh/commit/79deb6029f77c618f4383a092382ecd1733a30b7))
+* **apps/frontend-pwa:** visualization level in header ([#3399](https://github.com/uzh-bf/klicker-uzh/issues/3399)) ([d9a8214](https://github.com/uzh-bf/klicker-uzh/commit/d9a8214885ee08bd752eec9adf5cbe68fb11101b))
+* award xp during live sessions ([#3403](https://github.com/uzh-bf/klicker-uzh/issues/3403)) ([a2880ed](https://github.com/uzh-bf/klicker-uzh/commit/a2880ed283832e886ef39206cc85c6fa452fc27b))
+
+## [3.0.0-alpha.66](https://github.com/uzh-bf/klicker-uzh/compare/v3.0.0-alpha.65...v3.0.0-alpha.66) (2023-03-01)
+
+
+### Bug Fixes
+
+* **packages/shared-components:** ensure that options do not appear selected on mobile devices ([#3419](https://github.com/uzh-bf/klicker-uzh/issues/3419)) ([8ea7910](https://github.com/uzh-bf/klicker-uzh/commit/8ea7910ac86c8e8f2adfe3ebc87881d03c6e39aa))
+
+## [3.0.0-alpha.65](https://github.com/uzh-bf/klicker-uzh/compare/v3.0.0-alpha.64...v3.0.0-alpha.65) (2023-03-01)
+
+
+### Bug Fixes
+
+* **apps/frontend-manage:** fix pre-filling fields when editing micro sessions ([#3418](https://github.com/uzh-bf/klicker-uzh/issues/3418)) ([f7aad42](https://github.com/uzh-bf/klicker-uzh/commit/f7aad4223a43d2c4210f489f9c0fdb10df58ff3e))
+
+
+### Refactors
+
+* flex-based layout for PWA ([0be2fff](https://github.com/uzh-bf/klicker-uzh/commit/0be2fffd5f55774c093de23e3bf5c0542ee8279c))
+
 ## [3.0.0-alpha.64](https://github.com/uzh-bf/klicker-uzh/compare/v3.0.0-alpha.63...v3.0.0-alpha.64) (2023-02-27)